update jwt  authen

+++ src/f16/dru/dru.service.ts
@@ -4,7 +4,7 @@ import { PrismaFinance } from 'src/prisma/prisma.service.finanec';
 
 @Injectable()
 export class DruService {
-  constructor(readonly prisma: PrismaFinance) {}
+  constructor(readonly prisma: PrismaFinance) { }
 
   createDru = async (dru: t_dru[]) => {
     const listSize = 1000;
@@ -24,27 +24,45 @@ export class DruService {
   };
 
   findAndDelete = async (drulist: Array<t_dru[]>) => {
+    console.log('start dru delete');
+
     return await new Promise((resolve, reject) => {
       const apdListSzie = drulist.length;
       let round = 0;
       drulist.forEach(async (dru) => {
-        await this.prisma.t_dru
-          .deleteMany({
-            where: {
-              seq: { in: dru.map((i) => i.seq) },
-            },
-          })
-          .finally(() => this.prisma.$disconnect());
+        const seq: string[] = [];
+        const an: string[] = [];
 
-        await this.prisma.t_dru
-          .deleteMany({
-            where: {
-              an: { in: dru.map((i) => i.an) },
-            },
-          })
-          .finally(() => this.prisma.$disconnect());
+        dru.forEach((i) => {
+          if (i.an !== '') {
+            an.push(i.an);
+          }
+
+          if (i.seq !== '') {
+            seq.push(i.seq);
+          }
+        });
+
+        if (seq.length > 0) {
+          await this.prisma.t_dru
+            .deleteMany({
+              where: {
+                seq: { in: seq },
+              },
+            })
+        }
+
+        if (an.length > 0) {
+          await this.prisma.t_dru
+            .deleteMany({
+              where: {
+                an: { in: an },
+              },
+            })
+        }
 
         round = round + 1;
+        console.log('dru round=' + round.toString() + ' max=' + apdListSzie.toString());
 
         if (apdListSzie === round) {
           resolve(true);
@@ -55,19 +73,20 @@ export class DruService {
 
   insert = async (druList: Array<t_dru[]>) => {
     return await new Promise((resolve, reject) => {
+      console.log('start create dru');
+      console.log(druList.length);
+
       const listSize = druList.length;
       let round = 0;
       druList.forEach(async (dru) => {
-        await this.prisma.t_dru
-          .createMany({
-            data: dru,
-          })
-          .finally(() => this.prisma.$disconnect());
-
+        console.log(dru.length);
+        await this.prisma.t_dru.createMany({
+          data: dru,
+        })
         round = round + 1;
 
         if (listSize === round) {
-          resolve(true);
+          resolve(true)
         }
       });
     });
